Use DB as source of truth for onboarding enforcement
